Refactor and simplify initing a new agent project
